smart_energy_monitor_deployment/requirements.txt
# Core dependencies
flask>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
schedule>=1.1.0
scikit-learn>=1.0.0

# Web interface
jinja2>=3.0.0

# Data visualization
plotly>=5.0.0
matplotlib>=3.5.0

# Smart plug integration
python-kasa>=0.5.0  # TP-Link Kasa plugs

# Async operations
aiohttp>=3.8.0

# Hardware interface (Raspberry Pi - optional)
# Uncomment if running on Raspberry Pi:
# spidev>=3.5          # SPI interface for ADC
# RPi.GPIO>=0.7.0      # GPIO control
# gpiozero>=1.6.0      # Simplified GPIO interface

# AI integration (optional - for gpt-oss when available)
# Uncomment for AI features:
# transformers>=4.20.0
# torch>=1.12.0
# accelerate>=0.20.0